requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.3.1
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.3
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
numpy==2.0.0
openai==1.35.3
packaging==24.1
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.32.3
sniffio==1.3.1
soundfile==0.12.1
tomli==2.0.1
tqdm==4.66.4
twilio==9.2.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
